Refactor moar
