+++ RobotServer/controllers/http_robot/http_robot.py
@@ -11,10 +11,11 @@ import traceback
 from controller import Supervisor
 from flask import Flask, request
 
-from device_map import build_device_map
-from motor_requests import apply_motor_requests, update_motors
-from sensors import get_sensors_data, get_world_pose
-
+from robot.device_map import build_device_map
+from robot.motor_requests import apply_motor_requests, update_motors
+from robot.sensors import get_sensors_data, get_world_pose
+from webots_draw import draw_line, draw_arrow, draw_circle
+from camera_stream import start_zmq
 
 # flask log level
 log = logging.getLogger("werkzeug")
@@ -75,7 +76,7 @@ def create_app(device_map, motor_requests, robot):
         point_2 = request_data["point_2"]
         color_rgb = request_data.get("color", [0, 1, 1])
 
-        draw_line(name, point_1, point_2, color_rgb=color_rgb)
+        draw_line(robot, line_map, name, point_1, point_2, color_rgb=color_rgb)
 
         return "OK"
 
@@ -88,7 +89,7 @@ def create_app(device_map, motor_requests, robot):
         point_2 = request_data["point_2"]
         color_rgb = request_data.get("color", [0, 1, 1])
 
-        draw_arrow(name, point_1, point_2, color_rgb=color_rgb)
+        draw_arrow(robot, line_map, name, point_1, point_2, color_rgb=color_rgb)
 
         return "OK"
 
@@ -101,7 +102,14 @@ def create_app(device_map, motor_requests, robot):
         radius = request_data["radius"]
         color_rgb = request_data.get("color", [0, 1, 1])
 
-        draw_circle(name=name, center=center, color_rgb=color_rgb, radius=radius)
+        draw_circle(
+            robot,
+            line_map,
+            name=name,
+            center=center,
+            color_rgb=color_rgb,
+            radius=radius,
+        )
 
         return "OK"
 
@@ -121,199 +129,6 @@ def get_public_ip():
         return "127.0.0.1"
 
 
-# Draws a line between point_1 and point_2
-# name creates the node in the tree so we just update the same node each time
-def draw_line(
-    name: str, point_1: List[float], point_2: List[float], color_rgb: List[float]
-):
-    # Create node with name if it doesn't exist yet
-    node_name = f"LINE_{name}"
-
-    if node_name not in line_map:
-        root_node = robot.getRoot()
-        root_children_field = root_node.getField("children")
-
-        template = (
-            f"DEF {node_name} "
-            + """Shape {
-            appearance Appearance {
-                material Material {
-                    emissiveColor """
-            + " ".join(map(str, color_rgb))
-            + """
-                }
-            }
-            geometry DEF TRAIL_LINE_SET IndexedLineSet {
-                coord Coordinate {
-                    point [
-                        0 0 0
-                        0 0 0
-                    ]
-                }
-                coordIndex [
-                    0 1 -1
-                ]
-            }
-        }"""
-        )
-        root_children_field.importMFNodeFromString(-1, template)
-
-        # Update the coords based on points
-        trail_set_node = robot.getFromDef(f"{node_name}.TRAIL_LINE_SET")
-        coordinates_node = trail_set_node.getField("coord").getSFNode()
-        point_field = coordinates_node.getField("point")
-        line_map[node_name] = point_field
-    else:
-        point_field = line_map[node_name]
-
-    point_field.setMFVec3f(0, point_1)
-    point_field.setMFVec3f(1, point_2)
-
-
-# Draws a line between point_1 and point_2
-# name creates the node in the tree so we just update the same node each time
-def draw_arrow(
-    name: str,
-    point_1: List[float],
-    point_2: List[float],
-    color_rgb: List[float],
-    arrow_angle_deg=35,
-    arrow_len_ratio=0.25,
-):
-    # Create node with name if it doesn't exist yet
-    node_name = f"ARROW_{name}"
-
-    if node_name not in line_map:
-        root_node = robot.getRoot()
-        root_children_field = root_node.getField("children")
-
-        template = (
-            f"DEF {node_name} "
-            + """Shape {
-            appearance Appearance {
-                material Material {
-                    emissiveColor """
-            + " ".join(map(str, color_rgb))
-            + """
-                }
-            }
-            geometry DEF TRAIL_LINE_SET IndexedLineSet {
-                coord Coordinate {
-                    point [
-                        0 0 0
-                        0 0 0
-                        0 0 0
-                        0 0 0
-                    ]
-                }
-                coordIndex [
-                    0 1 -1
-                    1 2 -1
-                    1 3 -1
-                ]
-            }
-        }"""
-        )
-        root_children_field.importMFNodeFromString(-1, template)
-
-        # Update the coords based on points
-        trail_set_node = robot.getFromDef(f"{node_name}.TRAIL_LINE_SET")
-        coordinates_node = trail_set_node.getField("coord").getSFNode()
-        point_field = coordinates_node.getField("point")
-        line_map[node_name] = point_field
-    else:
-        point_field = line_map[node_name]
-
-    # calculate arrow points based on start/end
-    # convert to polar
-    delta = [point_1[0] - point_2[0], point_1[1] - point_2[1]]
-    length = math.sqrt(delta[0] ** 2 + delta[1] ** 2)
-    angle = math.atan2(delta[0], delta[1])
-
-    left_angle = angle + math.radians(arrow_angle_deg)
-    right_angle = angle - math.radians(arrow_angle_deg)
-
-    arrow_length = length * arrow_len_ratio
-    arrow_left = [
-        arrow_length * math.sin(left_angle) + point_2[0],
-        arrow_length * math.cos(left_angle) + point_2[1],
-        point_2[2],
-    ]
-    arrow_right = [
-        arrow_length * math.sin(right_angle) + point_2[0],
-        arrow_length * math.cos(right_angle) + point_2[1],
-        point_2[2],
-    ]
-
-    point_field.setMFVec3f(0, point_1)
-    point_field.setMFVec3f(1, point_2)
-    point_field.setMFVec3f(2, arrow_left)
-    point_field.setMFVec3f(3, arrow_right)
-
-
-def draw_circle(
-    name: str,
-    center: List[float],
-    color_rgb: List[float],
-    radius: float,
-    num_segments=20,
-):
-    # Create node with name if it doesn't exist yet
-    node_name = f"CIRCLE_{name}"
-
-    if node_name not in line_map:
-        node = robot.getFromDef(node_name)
-        root_node = robot.getRoot()
-        root_children_field = root_node.getField("children")
-
-        template = (
-            f"DEF {node_name} "
-            + """Shape {
-            appearance Appearance {
-                material Material {
-                    emissiveColor """
-            + " ".join(map(str, color_rgb))
-            + """
-                }
-            }
-            geometry DEF TRAIL_LINE_SET IndexedLineSet {
-                coord Coordinate {
-                    point [
-                        """
-            + ("0 0 0\n" * num_segments)
-            + """
-                    ]
-                }
-                coordIndex [
-                    """
-            + " ".join(map(str, range(num_segments)))
-            + """ 0 -1
-                ]
-            }
-        }"""
-        )
-        root_children_field.importMFNodeFromString(-1, template)
-        node = robot.getFromDef(node_name)
-
-        # Update the coords based on points
-        trail_set_node = robot.getFromDef(f"{node_name}.TRAIL_LINE_SET")
-        coordinates_node = trail_set_node.getField("coord").getSFNode()
-        point_field = coordinates_node.getField("point")
-        line_map[node_name] = point_field
-    else:
-        point_field = line_map[node_name]
-
-    # Use polar coordinates to find points on the circle
-    for i in range(num_segments):
-        theta = math.pi * 2 / num_segments * i
-        point = [
-            center[0] + radius * math.cos(theta),
-            center[1] + radius * math.sin(theta),
-            center[2],
-        ]
-        point_field.setMFVec3f(i, point)
-
-
 def start_flask(device_map, motor_requests, robot):
 
     flask_app = create_app(
@@ -329,51 +144,6 @@ def start_flask(device_map, motor_requests, robot):
     flask_app.run(host="0.0.0.0", port=port)
 
 
-def start_zmq():
-    global device_map, port
-
-    zmq_video_sleep_seconds = 1 / 60
-    zmq_port_offset = 10
-    zmq_port = port + zmq_port_offset
-    print(
-        f"[HttpRobot{robotId}] Starting zmq server on tcp://{get_public_ip()}:{zmq_port}",
-        flush=True,
-    )
-
-    try:
-        import zmq
-
-        # Start a ZeroMQ server.
-        zmq_context = zmq.Context()
-        zmq_socket = zmq_context.socket(zmq.PUB)
-        zmq_socket.bind(f"tcp://0.0.0.0:{zmq_port}")
-
-        # Get the camera, and define a message template to send the image.
-        if not device_map["Cameras"]:
-            print("No camera found, skipping zmq setup")
-            return
-        camera = next(iter(device_map["Cameras"].values()))
-        camera_depth = 4
-        message = [
-            b"image",
-            camera.getHeight().to_bytes(4, byteorder="big"),
-            camera.getWidth().to_bytes(4, byteorder="big"),
-            camera_depth.to_bytes(4, byteorder="big"),
-            None,
-        ]
-
-        while True:
-            # Get the camera image and send the full message.
-            message[-1] = bytes(camera.getImage())
-            zmq_socket.send_multipart(message)
-
-            # Sleep so we don't overload the ZMQ socket.
-            # The sleep time is correlated with framerate, but doesn't seem to match it exactly.
-            time.sleep(zmq_video_sleep_seconds)
-
-    except:
-        print("Could not start zmq!", file=sys.stderr)
-        print(traceback.format_exc(), file=sys.stderr)
 
 
 if __name__ == "__main__":
@@ -389,7 +159,7 @@ if __name__ == "__main__":
     threading.Thread(
         target=lambda: start_flask(device_map, motor_requests, robot)
     ).start()
-    threading.Thread(target=start_zmq).start()
+    threading.Thread(target=lambda: start_zmq(port, robotId, get_public_ip(), device_map)).start()
 
     # Run the simulation loop
     print("Starting null op simulation loop")
