+++ RobotServer/controllers/http_robot/robot/device_map.py
